requirements.txt
Flask==3.1.0
google-generativeai==0.8.3
pandas==2.2.3
matplotlib==3.9.3
seaborn==0.13.2
python-dotenv==1.0.1
